requirements.txt
python-dotenv
streamlit 
azure-ai-translation-text==1.0.0b1